+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * added `stubbles\values\Value::isEmpty()`
   * added possibility to enforce strict type check in `stubbles\values\Value::isOneOf()`
   * added possibility to use a different character as separator for list values in `stubbles\values\Parse::toList()` and `stubbles\values\Parse::asList()`
+  * added `stubbles\values\Rootpath::default()`
 
 
 8.0.0 (2016-07-11)